refactor

+++ Src/Application/Hotels/Dtos/HotelDto.cs
@@ -5,8 +5,7 @@ public class HotelDto
     public Guid Id { get; set; }
     public string Name { get; set; } = string.Empty;
     public string Address { get; set; } = string.Empty;
-
-    public float Rating { get; set; }
+    public decimal Rating { get; set; }
     // public ICollection<ManagerDto> Managers { get; set; } = [];
     // public ICollection<RoomDto> Rooms { get; set; } = [];
 }
